// src/app/layout/models/userSetting.ts
export interface userSettings {
    ccId: number;
    columnId: string;
    columnName: string;
    dataType: string;
    isGolden: boolean;
    isDedupe: boolean;
    isHidden: boolean;
    usId: number;
    userEmailId: string;
    customerColumnId: number;
    defaultPageSize: string;
    usColumnId: string;
    isAssigned: boolean;
}